requirements.txt
pandas==2.2.2
numpy==2.0.0
matplotlib==3.9.0
pillow==10.3.0